Updated how Quartus workflow is run
Quartus workflow is now run as a system call in the background and its
log file is read in Matlab until exit is encountered

Additionally less html reports are generated by HDLCoder and it was
updated for the new target names. Additionally the logging module is now
used by the Quartus workflow.

Also fixes were made regarding updating the target from De10 to
Audiomini as compilitation previously was failing from the name change