Variety of updates to support the Beamer
Small fixes to parsing and style updates to dtogen code

Fix addr_width for avalon wrapper when there are no registers
Update channel to sample and sample to channel process generation to use audio in and out channel count
Updated channel to sample process generation to ensure the last channel isn't dropped

Updated num_to_bitstring to have what it surrounds the bitstring with to be configurable.
Ex. '0' or "0" or nothing with just 0

Updated max channel in the hw_tcl ip core to be set correctly

Updated vhdl generation to get the next sample time if the smallest is 0

Fixed tcl script generation used to generate qsys file for when there are multiple devices
Updated project creation to make a new revision instead of overwrite the project if a project with the same name exists
When compiling, it will force open the project which allows it to open the project even if it was last compiled for a different version of Quartus
Previous support for the Audioblade was actually a mix of the Audioblade and Reflex SoM so it has been transitioned to Reflex SoM support
Added support for generating jic files when a project is building for the Reflex SoM
Added support for target's having files in their file list copied over during project generation